Updated convertDateRangeFormat to use "short" months. Updated columns widths for Resume component. Adjusted vertical padding to image in Resume component.
 Changes to be committed:
	modified:   src/components/Home/components/Resume/Resume.tsx
	modified:   src/utils/convertDateRangeFormat.ts

+++ src/utils/convertDateRangeFormat.ts
@@ -11,7 +11,7 @@ function convertDateRangeFormat(dateRange: string) {
 function formatDate(dateString: string) {
   const [month, year] = dateString.split("/");
   const dateObj = new Date(parseInt(year), parseInt(month.slice(0)));
-  const fullMonth = dateObj.toLocaleString("en-US", { month: "long" });
+  const fullMonth = dateObj.toLocaleString("en-US", { month: "short" });
   return `${fullMonth} ${year}`;
 }
